# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yeshshree Operations API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/stock/balances": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Stock Balances",
        "operationId": "stock_balances_api_v1_stock_balances_get",
        "parameters": [
          {
            "name": "material_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Material Id"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "name": "vendor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vendor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BalanceRead"
                  },
                  "title": "Response Stock Balances Api V1 Stock Balances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stock/days-of-cover/{material_id}": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Stock Days Of Cover",
        "operationId": "stock_days_of_cover_api_v1_stock_days_of_cover__material_id__get",
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Material Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaysOfCoverRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/sap-stock": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Import Sap Stock",
        "description": "SAP stock snapshot \u2192 import_jobs(kind='sap_stock') + ADJUST reconcile (\u00a711.1).",
        "operationId": "import_sap_stock_api_v1_imports_sap_stock_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_sap_stock_api_v1_imports_sap_stock_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SapStockImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Post Issue",
        "description": "200 (not 201): a retried client_ref returns the ORIGINAL result (invariant 4).",
        "operationId": "post_issue_api_v1_issues_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "List Issues",
        "operationId": "list_issues_api_v1_issues_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "vendor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vendor Id"
            }
          },
          {
            "name": "destination",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Destination"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IssueRead"
                  },
                  "title": "Response List Issues Api V1 Issues Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/{issue_id}": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Get Issue",
        "operationId": "get_issue_api_v1_issues__issue_id__get",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/{issue_id}/correction": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Request Correction",
        "operationId": "request_correction_api_v1_issues__issue_id__correction_post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__inventory__CorrectionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__inventory__CorrectionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vendors/{vendor_id}/exposure": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Vendor Exposure",
        "operationId": "vendor_exposure_api_v1_vendors__vendor_id__exposure_get",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vendor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExposureRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List Notifications",
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "name": "unread",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unread"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationRead"
                  },
                  "title": "Response List Notifications Api V1 Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark Read",
        "operationId": "mark_read_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}/act": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark Acted",
        "description": "Sets acted_at \u2014 stops blocker repeats for the whole chain on this ref.",
        "operationId": "mark_acted_api_v1_notifications__notification_id__act_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalanceRead": {
        "properties": {
          "material_id": {
            "type": "integer",
            "title": "Material Id"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "vendor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          },
          "qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "material_id",
          "location",
          "qty"
        ],
        "title": "BalanceRead"
      },
...
      "Body_import_sap_stock_api_v1_imports_sap_stock_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_sap_stock_api_v1_imports_sap_stock_post"
      },
...
      "DaysOfCoverRead": {
        "properties": {
          "material_id": {
            "type": "integer",
            "title": "Material Id"
          },
          "days_of_cover": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Of Cover"
          }
        },
        "type": "object",
        "required": [
          "material_id",
          "days_of_cover"
        ],
        "title": "DaysOfCoverRead"
      },
...
      "ExposureRead": {
        "properties": {
          "vendor_id": {
            "type": "integer",
            "title": "Vendor Id"
          },
          "credit_exposure": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credit Exposure"
          },
          "qty_mt": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty Mt"
          },
          "credit_limit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Limit"
          },
          "qty_limit_mt": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qty Limit Mt"
          }
        },
        "type": "object",
        "required": [
          "vendor_id",
          "credit_exposure",
          "qty_mt",
          "credit_limit",
          "qty_limit_mt"
        ],
        "title": "ExposureRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IssueCreate": {
        "properties": {
          "destination": {
            "type": "string",
            "enum": [
              "inhouse",
              "vendor_sale",
              "job_work"
            ],
            "title": "Destination"
          },
          "material_id": {
            "type": "integer",
            "title": "Material Id"
          },
          "qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Qty"
          },
          "uom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uom"
          },
          "line_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Id"
          },
          "vendor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          },
          "client_ref": {
            "type": "string",
            "format": "uuid",
            "title": "Client Ref"
          }
        },
        "type": "object",
        "required": [
          "destination",
          "material_id",
          "qty",
          "client_ref"
        ],
        "title": "IssueCreate"
      },
      "IssueRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "doc_no": {
            "type": "string",
            "title": "Doc No"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "line_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Id"
          },
          "vendor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          },
          "material_id": {
            "type": "integer",
            "title": "Material Id"
          },
          "qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty"
          },
          "uom": {
            "type": "string",
            "title": "Uom"
          },
          "value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Value"
          },
          "limit_check": {
            "additionalProperties": true,
            "type": "object",
            "title": "Limit Check"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "stock_warning": {
            "type": "boolean",
            "title": "Stock Warning",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "doc_no",
          "destination",
          "line_id",
          "vendor_id",
          "material_id",
          "qty",
          "uom",
          "value",
          "limit_check",
          "status",
          "approval_id",
          "created_by",
          "created_at"
        ],
        "title": "IssueRead"
      },
...
      "NotificationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "ref_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Type"
          },
          "ref_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Id"
          },
          "read_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "acted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acted At"
          },
          "next_repeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Repeat At"
          },
          "digest_of": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digest Of"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "body",
          "kind",
          "ref_type",
          "ref_id",
          "read_at",
          "created_at",
          "tier",
          "acted_at",
          "next_repeat_at",
          "digest_of"
        ],
        "title": "NotificationRead"
      },
...
      "SapStockImportResult": {
        "properties": {
          "import_job_id": {
            "type": "integer",
            "title": "Import Job Id"
          },
          "rows_total": {
            "type": "integer",
            "title": "Rows Total"
          },
          "rows_ok": {
            "type": "integer",
            "title": "Rows Ok"
          },
          "rows_failed": {
            "type": "integer",
            "title": "Rows Failed"
          },
          "adjustments": {
            "type": "integer",
            "title": "Adjustments"
          }
        },
        "type": "object",
        "required": [
          "import_job_id",
          "rows_total",
          "rows_ok",
          "rows_failed",
          "adjustments"
        ],
        "title": "SapStockImportResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__inventory__CorrectionCreate": {
        "properties": {
          "corrected": {
            "$ref": "#/components/schemas/CorrectedFields"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "corrected",
          "reason"
        ],
        "title": "CorrectionCreate"
      },
      "app__schemas__inventory__CorrectionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "issue_id": {
            "type": "integer",
            "title": "Issue Id"
          },
          "corrected": {
            "additionalProperties": true,
            "type": "object",
            "title": "Corrected"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "approval_id": {
            "type": "integer",
            "title": "Approval Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "issue_id",
          "corrected",
          "reason",
          "approval_id",
          "status"
        ],
        "title": "CorrectionRead"
      },
...
    }
  }
}